@@ -1,6 +1,6 @@
 {
   "name": "venecodollar",
-  "version": "2.2.2",
+  "version": "2.3.0",
   "description": "Library to consult the different dollar and euro values and their average in terms of bolivars. The data is obtained from https://monitordolarvenezuela.com/",
   "keywords": [
     "dollar",
@@ -17,9 +17,17 @@
     "deploy-build": "ncc build src/index.ts -o public -m",
     "package-build": "rimraf ./build && tsc --project package.tsconfig.json",
     "build": "npm run deploy-build & npm run package-build",
-    "build:start": "npm run deploy-build & npm run package-build & @powershell -Command $env:DEBUG='bot*';dotenv -- -- nodemon -e ts -x ts-node src/index.ts",
-    "prettier": "prettier --write 'src/**/*.ts'",
-    "package-deploy": "npm publish --access public"
+    "build:start": "npm run deploy-build & npm run package-build & npm run devWindows",
+    "release": "npm run build && changeset publish",
+    "update:version": "changeset",
+    "publish": "npm publish --access=public",
+    "test": "jest",
+    "lint": "eslint src/**/*.ts",
+    "lint:fix": "eslint src/**/*.ts --fix",
+    "format:check": "prettier . --check",
+    "format:write": "prettier . --write",
+    "force:check": "eslint src/**/*.ts & prettier . --check",
+    "force:fix": "eslint src/**/*.ts --fix & prettier . --write"
   },
   "author": "gustavoerivero",
   "license": "MIT",
@@ -34,27 +42,42 @@
     "cheerio": "^1.0.0-rc.12"
   },
   "devDependencies": {
+    "@babel/core": "^7.25.8",
+    "@babel/preset-env": "^7.25.8",
+    "@babel/preset-typescript": "^7.25.7",
+    "@changesets/cli": "^2.27.9",
     "@types/cors": "^2.8.13",
     "@types/express": "^4.17.17",
+    "@types/jest": "^29.5.13",
     "@types/morgan": "^1.9.4",
-    "@types/node": "^20.2.6",
+    "@types/node": "^22.7.5",
     "@types/swagger-jsdoc": "^6.0.1",
-    "@types/swagger-ui-express": "^4.1.3",
+    "@types/swagger-ui-express": "^4.1.6",
+    "@typescript-eslint/eslint-plugin": "^8.8.1",
+    "@typescript-eslint/parser": "^8.8.1",
     "@vercel/ncc": "^0.36.1",
     "axios": "^1.5.1",
+    "babel-jest": "^29.7.0",
     "cors": "^2.8.5",
     "debug": "^4.3.4",
     "dotenv": "^16.1.4",
     "dotenv-cli": "^7.2.1",
+    "eslint": "^9.12.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-jest": "^28.8.3",
+    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-unused-imports": "^4.1.4",
     "express": "^4.18.2",
-    "jest": "^29.5.0",
+    "jest": "^29.7.0",
     "morgan": "^1.10.0",
     "nodemon": "^2.0.22",
     "pre-commit": "^1.2.2",
+    "prettier": "^3.3.3",
     "rimraf": "^5.0.1",
     "swagger-jsdoc": "^6.2.8",
     "swagger-ui-express": "^4.6.3",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.1.3"
+    "ts-jest": "^29.2.5",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.6.3"
   }
 }
