@@ -1,2 +1,2 @@
-export * from "./CDollar"
-export * from "./CEuro"
+export * from './CDollar';
+export * from './CEuro';
