@@ -1,40 +1,42 @@
 // Dependencies
-import express, { Response } from "express"
-import cors from "cors"
-import morgan from "morgan"
-import routes from "./routes"
-import swaggerDocs from "./services/swagger"
-import swaggerUI from "swagger-ui-express"
+import express, { Response } from 'express';
+import cors from 'cors';
+import morgan from 'morgan';
+import routes from './routes';
+import swaggerDocs from './services/swagger';
+import swaggerUI from 'swagger-ui-express';
 
 // Dotenv
-require("dotenv").config()
+import dotenv from 'dotenv';
+dotenv.config();
 
 // Settings
-const port = process.env.PORT ?? 3030
-const app = express()
+const port = process.env.PORT ?? 3030;
+const app = express();
 
 // Middlewares
-app.use(express.json())
-app.use(morgan("tiny"))
-app.use(express.urlencoded({ extended: false }))
+app.use(express.json());
+app.use(morgan('tiny'));
+app.use(express.urlencoded({ extended: false }));
 
-app.use(cors())
+app.use(cors());
 
 // Routes
-const URL = `${process.env.API_VER_URL}`
-const OLD = `${process.env.API_OLD}`
-const DOC_URL = `${process.env.API_DOC}`
+const URL = `${process.env.API_VER_URL}`;
+const OLD = `${process.env.API_OLD}`;
+const DOC_URL = `${process.env.API_DOC}`;
 
-app.use(URL, routes)
-app.use(OLD, routes)
-app.use(DOC_URL, swaggerUI.serve, swaggerUI.setup(swaggerDocs))
+app.use(URL, routes);
+app.use(OLD, routes);
+app.use(DOC_URL, swaggerUI.serve, swaggerUI.setup(swaggerDocs));
 
 // Start server
-app.get(URL, (_, res: Response) => res.send(`Connected!`))
+app.get(URL, (_, res: Response) => res.send(`Connected!`));
+app.get(OLD, (_, res: Response) => res.send(`Connected on old version!`));
 
 app.listen(port, () => {
-  console.log(`Server is running on ${URL}`)
-})
+  console.log(`Server is running on ${URL}`);
+});
 
 /**
  * This module is responsible for setting up an Express server that listens for incoming requests.
