@@ -1,9 +1,9 @@
-declare module "*.json" {
-  const value: any
-  export default value
+declare module '*.json' {
+  const value: any;
+  export default value;
 }
 
-declare module "*.yml" {
-  const value: any
-  export default value
+declare module '*.yml' {
+  const value: any;
+  export default value;
 }
