@@ -1,8 +1,9 @@
 //import * as cheerio from "cheerio"
-import { TBsDollarCalculated, TDollar, TDollarAverage, TDollarCalculated, TDollarEntity } from "../types/TDollar"
-import { formatDate } from "../utils/formatDate"
-import http from "./http"
-import { TResponse } from "../types";
+import { TBsDollarCalculated, TDollar, TDollarAverage, TDollarCalculated, TDollarEntity } from '../types/TDollar';
+import { formatDate } from '../utils/formatDate';
+import http from './http';
+import { TResponse } from '../types';
+import { AxiosResponse } from 'axios';
 
 const ABS = 5;
 
@@ -10,15 +11,17 @@ const decimals = (number: number, decimals: number = 2) => Number(number.toFixed
 
 const getTendency = (tendency: number) => {
   if (tendency > 0) {
-    return "Uptrend";
+    return 'Uptrend';
   } else if (tendency < 0) {
-    return "Downtrend"
+    return 'Downtrend';
   } else {
-    return "Unchanged";
+    return 'Unchanged';
   }
 };
 
-const discard = (pivot: number, minus: number = 0, abs: number = 0) => ((Math.abs(pivot - minus) - abs) > abs);
+const discard = (pivot: number, minus: number = 0, abs: number = 0) => Math.abs(pivot - minus) - abs > abs;
+const getEntity = (data: AxiosResponse<TResponse[], any>, name: string) =>
+  data.data.find((item) => item.name.toLowerCase().includes(name));
 
 /**
  * Fetches an array with different values of the dollar in bolivars managed by entities that monitor this value.
@@ -31,44 +34,47 @@ export const getDollarPrices = async (): Promise<TDollar[] | null> => {
   try {
     // Fetch data from the specified URL
 
-    const EXT = "/coins"
+    const EXT = '/coins';
     const api = http();
 
     const response = await api.get<TResponse[]>(EXT);
 
     if (!response) {
-      throw new Error("Request failed")
+      throw new Error('Request failed');
     }
 
     // Parse text response from fetch function.
     let { data } = response;
     const pivot = data[0];
-    const petro = data.find(item => item.name.toLowerCase().includes("petro"));
+    const bcv = getEntity(response, 'bcv');
+    const petro = getEntity(response, 'petro');
 
-    data = data.filter(item => item.currency === "VES" && !discard(item.price, pivot.price, ABS));
+    data = data.filter(
+      (item) => item.currency === 'VES' && !discard(item.price, pivot.price, ABS) && item !== bcv && item !== petro,
+    );
+    bcv && data.unshift(bcv);
     petro && data.push(petro);
 
-    const priceResult: TDollar[] = []
+    const priceResult: TDollar[] = [];
 
-    data.forEach(item => {
-
+    data.forEach((item) => {
       let { name, price, icon, updatedAt, price24h } = item;
 
       const difference = decimals(price - (price24h ?? 0));
       const percentage = decimals(difference / price).toString();
 
-      let color = "Unchanged";
+      let color = 'Unchanged';
       const tendency = getTendency(difference);
 
-      if (tendency === "Uptrend") {
-        color = "green";
-      } 
-      
-      if (tendency === "Downtrend") {
-        color = "red";
+      if (tendency === 'Uptrend') {
+        color = 'green';
       }
 
-      name = name.replace("\u00F3", "ó")
+      if (tendency === 'Downtrend') {
+        color = 'red';
+      }
+
+      name = name.replace('\u00F3', 'ó');
 
       const dollarData: TDollar = {
         title: name,
@@ -78,25 +84,24 @@ export const getDollarPrices = async (): Promise<TDollar[] | null> => {
         difference: Number(Number(difference ?? 0).toFixed(2)),
         differencePercentage: percentage,
         tendency: tendency,
-        tendencyColor: color
-      }
+        tendencyColor: color,
+      };
 
       priceResult.push(dollarData);
-
     });
 
     // Return the array of dollar values
-    return priceResult
+    return priceResult;
   } catch (error) {
     // Handle error obtaining dollar values
-    console.error(`Error obtaining dollar values.`, error)
+    console.error(`Error obtaining dollar values.`, error);
     // Return null if an error occurs
-    return null
+    return null;
   }
-}
+};
 
 /**
- * Fetches an array with different values of the dollar in bolivars managed by entities that monitor this value. 
+ * Fetches an array with different values of the dollar in bolivars managed by entities that monitor this value.
  * It also calculates the average of all entities with values greater than zero.
  *
  * @returns {Promise<TAverage | null>} - A promise that resolves to an array with different dollar values
@@ -106,51 +111,50 @@ export const getDollarPrices = async (): Promise<TDollar[] | null> => {
 export const getDollarPricesWithAverage = async (): Promise<TDollarAverage | null> => {
   try {
     // Fetch dollar prices from a remote source
-    const priceResult: TDollar[] | null = await getDollarPrices()
+    const priceResult: TDollar[] | null = await getDollarPrices();
 
     if (priceResult) {
-      let average = 0
-      let length = 0
+      let average = 0;
+      let length = 0;
 
       // Calculate average and create entities array
       const prices = priceResult.map((price: TDollar) => {
-
         const name = price.title.toLowerCase();
 
-        average = name !== "petro" && name !== "petro bs" ? Number(average) + Number(price.dollar) : Number(average)
-        length = Number(price.dollar) > 0 && name !== "petro" && name !== "petro bs" ? length + 1 : length
+        average = name !== 'petro' && name !== 'petro bs' ? Number(average) + Number(price.dollar) : Number(average);
+        length = Number(price.dollar) > 0 && name !== 'petro' && name !== 'petro bs' ? length + 1 : length;
 
-        let entity: TDollarEntity = {
+        const entity: TDollarEntity = {
           entity: price.title,
-          info: price
-        }
+          info: price,
+        };
 
-        return entity
-      })
+        return entity;
+      });
 
       // Create response object with average and entities array
       const response: TDollarAverage = {
         date: new Date(),
         average: Number((average / length).toFixed(2)),
-        entities: prices
-      }
+        entities: prices,
+      };
 
       // Return the response object
-      return response
+      return response;
     }
 
     // Return null if priceResult is null
-    return null
+    return null;
   } catch (error) {
     // Handle error calculating data
-    console.error(`Error calculating data.`, error)
+    console.error(`Error calculating data.`, error);
     // Return null if an error occurs
-    return null
+    return null;
   }
-}
+};
 
 /**
-* Fetches an array with the different values of the dollar in bolivars handled by the entities that control this value and calculates the value of the amount of dollars supplied in bolivars. 
+ * Fetches an array with the different values of the dollar in bolivars handled by the entities that control this value and calculates the value of the amount of dollars supplied in bolivars.
  * @param dollar {number} - Amount in dollars to be calculated in bolivars.
  * @returns {Promise<TBsCalculated[] | null>} - A promise that resolves to an array with different dollar values.
  * in bolivars handled by entities that control this value, along with the calculation in bolivars of the amount supplied in dollars as a parameter. Returns null if an error occurs.
@@ -158,36 +162,37 @@ export const getDollarPricesWithAverage = async (): Promise<TDollarAverage | nul
  */
 export const calculateDollarToBs = async (dollar: number): Promise<TBsDollarCalculated[] | null> => {
   try {
-
     if (!dollar || dollar <= 0) {
-      return null
+      return null;
     }
 
-    const entities = await getDollarPricesWithAverage()
+    const entities = await getDollarPricesWithAverage();
 
-    let calculatedEntities: TBsDollarCalculated[] = []
+    const calculatedEntities: TBsDollarCalculated[] = [];
 
     if (entities?.entities && entities?.entities.length > 0) {
       entities.entities.forEach((item) => {
         calculatedEntities.push({
           ...item,
-          bolivarCalculated: Number(item.info.dollar) > 0 && item.info.title !== "Petro" ? Number(Number(Number(item.info.dollar) * dollar).toFixed(2)) : 0
-        })
-      })
+          bolivarCalculated:
+            Number(item.info.dollar) > 0 && item.info.title !== 'Petro'
+              ? Number(Number(Number(item.info.dollar) * dollar).toFixed(2))
+              : 0,
+        });
+      });
     }
 
-    return calculatedEntities
-
+    return calculatedEntities;
   } catch (error) {
     // Handle error calculating data
-    console.error(`Error calculating data.`, error)
+    console.error(`Error calculating data.`, error);
     // Return null if an error occurs
-    return null
+    return null;
   }
-}
+};
 
 /**
-* Fetches an array with the different values of the bolivars in dollars handled by the entities that control this value and calculates the value of the amount of bolivars supplied in dollars. 
+ * Fetches an array with the different values of the bolivars in dollars handled by the entities that control this value and calculates the value of the amount of bolivars supplied in dollars.
  * @param bs {number} - Amount in bolivars to be calculated in dollars.
  * @returns {Promise<TDollarCalculated[] | null>} - A promise that resolves to an array with different dollar values.
  * in bolivars handled by entities that control this value, along with the calculation in bolivars of the amount supplied in dollars as a parameter. Returns null if an error occurs.
@@ -195,30 +200,31 @@ export const calculateDollarToBs = async (dollar: number): Promise<TBsDollarCalc
  */
 export const calculateBsToDollar = async (bs: number): Promise<TDollarCalculated[] | null> => {
   try {
-
     if (!bs || bs <= 0) {
-      return null
+      return null;
     }
 
-    const entities = await getDollarPricesWithAverage()
+    const entities = await getDollarPricesWithAverage();
 
-    let calculatedEntities: TDollarCalculated[] = []
+    const calculatedEntities: TDollarCalculated[] = [];
 
     if (entities?.entities && entities?.entities.length > 0) {
-      entities.entities.forEach(item => {
+      entities.entities.forEach((item) => {
         calculatedEntities.push({
           ...item,
-          dollarCalculated: Number(item.info.dollar) > 0 && item.info.title !== "Petro" ? Number(Number(bs / Number(item.info.dollar)).toFixed(2)) : 0
-        })
-      })
+          dollarCalculated:
+            Number(item.info.dollar) > 0 && item.info.title !== 'Petro'
+              ? Number(Number(bs / Number(item.info.dollar)).toFixed(2))
+              : 0,
+        });
+      });
     }
 
-    return calculatedEntities
-
+    return calculatedEntities;
   } catch (error) {
     // Handle error calculating data
-    console.error(`Error calculating data.`, error)
+    console.error(`Error calculating data.`, error);
     // Return null if an error occurs
-    return null
+    return null;
   }
-}
+};
