@@ -2,7 +2,7 @@
  * Base URL.
  * export const BASE_URL = process.env.BASE_URL ?? "";
  */
-export const BASE_URL = process.env.SERVICE_API ?? "";
+export const BASE_URL = process.env.SERVICE_API ?? '';
 
-export * from './dollar'
-export * from './euro'
+export * from './dollar';
+export * from './euro';
