@@ -1,4 +1,4 @@
-export * from "./TDollar";
-export * from "./TEuro";
-export * from "./TTimeUnit";
-export * from "./TResponse";
+export * from './TDollar';
+export * from './TEuro';
+export * from './TTimeUnit';
+export * from './TResponse';
